do not rewrite config if exists

+++ docker/algod-relay/mainnet/context/run.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
-
-cp mainnet/* data/ -R
+if [ ! -f data/config.json ]
+then
+    cp mainnet/* data/ -R
+fi
 diagcfg telemetry enable
 diagcfg metric enable
 goal node start
